+++ src/web/main.cpp
@@ -1,3 +1,9 @@
 #include <emscripten.h>
 
-int EMSCRIPTEN_KEEPALIVE main() { return 0; }
+#include "Logging.h"
+
+int EMSCRIPTEN_KEEPALIVE main() {
+    logging::disable();
+
+    return 0;
+}